REFACT :: resolve conflict
