+++ src/presentation/auth/auth.controller.ts
@@ -8,9 +8,13 @@ import {
   Inject,
   Post,
   Res,
+  Query,
 } from '@nestjs/common';
 import { Response } from 'express';
-import { GenerateKakaoTokenUseCase, SignupUseCase } from '../../core/auth/port/auth.in.port';
+import {
+  GenerateKakaoTokenUseCase,
+  SignupUseCase,
+} from '../../core/auth/port/auth.in.port';
 import { SignupRequest } from './dto/auth.request';
 
 @Controller('auth')
@@ -37,12 +41,11 @@ export class AuthController {
 
   @Post('/kakao/signup') // API-Auth-202
   async kakaoSignup(
-    @Query('access_token') token: string,
+    @Headers('OA-TOKEN') token: string,
     @Body() req: SignupRequest,
     @Res() res: Response,
   ): Promise<Response> {
-    if (!token)
-      throw new HttpException('access_token: null일 수 없습니다.', 400);
+    if (!token) throw new BadRequestException('OA-TOKEN: null일 수 없습니다.');
 
     await this.signupUseCase.signupWithKakao(token, req);
